+++ R/RcppExports.R
@@ -2081,10 +2081,6 @@ Patch___FF16__FF16_Env__compute_environment <- function(obj_) {
     invisible(.Call('_plant_Patch___FF16__FF16_Env__compute_environment', PACKAGE = 'plant', obj_))
 }
 
-Patch___FF16__FF16_Env__compute_rates <- function(obj_) {
-    invisible(.Call('_plant_Patch___FF16__FF16_Env__compute_rates', PACKAGE = 'plant', obj_))
-}
-
 Patch___FF16__FF16_Env__reset <- function(obj_) {
     invisible(.Call('_plant_Patch___FF16__FF16_Env__reset', PACKAGE = 'plant', obj_))
 }
@@ -2203,10 +2199,6 @@ Patch___TF24__TF24_Env__introduce_new_node <- function(obj_, species_index) {
 
 Patch___TF24__TF24_Env__compute_environment <- function(obj_) {
     invisible(.Call('_plant_Patch___TF24__TF24_Env__compute_environment', PACKAGE = 'plant', obj_))
-}
-
-Patch___TF24__TF24_Env__compute_rates <- function(obj_) {
-    invisible(.Call('_plant_Patch___TF24__TF24_Env__compute_rates', PACKAGE = 'plant', obj_))
 }
 
 Patch___TF24__TF24_Env__reset <- function(obj_) {
@@ -2329,10 +2321,6 @@ Patch___TF24f__TF24_Env__compute_environment <- function(obj_) {
     invisible(.Call('_plant_Patch___TF24f__TF24_Env__compute_environment', PACKAGE = 'plant', obj_))
 }
 
-Patch___TF24f__TF24_Env__compute_rates <- function(obj_) {
-    invisible(.Call('_plant_Patch___TF24f__TF24_Env__compute_rates', PACKAGE = 'plant', obj_))
-}
-
 Patch___TF24f__TF24_Env__reset <- function(obj_) {
     invisible(.Call('_plant_Patch___TF24f__TF24_Env__reset', PACKAGE = 'plant', obj_))
 }
@@ -2451,10 +2439,6 @@ Patch___K93__K93_Env__introduce_new_node <- function(obj_, species_index) {
 
 Patch___K93__K93_Env__compute_environment <- function(obj_) {
     invisible(.Call('_plant_Patch___K93__K93_Env__compute_environment', PACKAGE = 'plant', obj_))
-}
-
-Patch___K93__K93_Env__compute_rates <- function(obj_) {
-    invisible(.Call('_plant_Patch___K93__K93_Env__compute_rates', PACKAGE = 'plant', obj_))
 }
 
 Patch___K93__K93_Env__reset <- function(obj_) {